requirements.txt
selenium>=4.15.0
pdfplumber>=0.10.0
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.0.0
webdriver-manager>=4.0.0
